checks if modgit was installed already

+++ bootstrap.sh
@@ -117,12 +117,16 @@ provision_env(){
 init_modgit(){
     #make app folder
     [ -d /var/app ] || mkdir -p /var/app
-
-    #ensure the deployment bed
-    [ -d /src/deployment ] || mkdir -p /src/deployment
-    curl https://raw.github.com/jeremyBass/modgit/master/modgit > /src/deployment/modgit
-    ln -s /src/deployment/modgit /usr/local/bin/modgit
-    chmod a=r+w+x /usr/local/bin/modgit
+    if [ -f /usr/local/bin/modgit ]
+    then
+        echo "modgit was already loaded"
+    else
+        #ensure the deployment bed
+        [ -d /src/deployment ] || mkdir -p /src/deployment
+        curl https://raw.github.com/jeremyBass/modgit/master/modgit > /src/deployment/modgit
+        ln -s /src/deployment/modgit /usr/local/bin/modgit
+        chmod a=r+w+x /usr/local/bin/modgit
+    if
     cd /var/app
     modgit init
 }
